physic_simulator: fix + change for double + add test

+++ SimpleSimulator/SimpleSimulator/Model/AquitisionCommunication/Aquisition.cs
@@ -51,7 +51,7 @@ namespace AquitisionCommunication{
         /// @param float latitude 
         /// @param float longitude
         /// </summary>
-        public void sentPosition(int id, float latitude, float longitude) {
+        public void sentPosition(int id, double latitude, double longitude) {
             // TODO implement here
         }
